+++ src/bika/lims/validators.py
@@ -448,11 +448,11 @@ class InterimFieldsValidator:
     def validate_choices(self, interim):
         """Checks whether the choices are valid for the given interim
         """
+        result_type = interim.get("result_type", "")
         choices = interim.get("choices")
         if not choices:
-            # No choices set, result type should remain empty
-            result_type = interim.get("result_type")
-            if not result_type:
+            # No choices set, result type should remain empty or multivalue
+            if result_type in ["", "multivalue"]:
                 return
             return _t(_("Control type is not supported for empty choices"))
 
@@ -480,6 +480,11 @@ class InterimFieldsValidator:
         if len(keys) < 2:
             return _t(_("At least, two options for choices field are required"))
 
+        # Multivalue is not supported with choices
+        if result_type in ["multivalue"]:
+            return _t(_(
+                "Multiple values control type is not supported for choices"
+            ))
 
 validation.register(InterimFieldsValidator())